refactor for PR #2676
